R2SA-backend/requirements.txt
boto3==1.24.28
botocore==1.27.59
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
colorama==0.4.6
colorlog==5.0.1
cryptography==39.0.1
idna==3.4
jmespath==0.10.0
pip==23.0.1
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
s3transfer==0.6.0
setuptools==66.0.0
six==1.16.0
urllib3==1.26.15
wheel==0.38.4
win-inet-pton==1.1.0

asgiref==3.6.0
brotlipy==0.7.0
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.14.0
et-xmlfile==1.0.1
gunicorn==20.1.0
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
psycopg2==2.9.3
pytz==2022.7
redis==3.5.3
celery==5.2.7
requests==2.28.2
sqlparse==0.4.3
stripe==2.49.0
typing_extensions==4.4.0
whitenoise==6.3.0
